Move utils/dom-based-weakref to core/dom/weakref
