+++ src/core/dom/weakref.js
